desktop-ui-live/live-ingest/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
kafka-python==2.0.2
python-multipart==0.0.20
